Snackbar работает - при нажатии отмены заметка восстанавливается, диалог на закрытие несохраненной заметки тоже работает (но пока все баги не убрал). Пока отключил разворот экрана для приложения
