+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,8 @@
         android:roundIcon="@mipmap/ic_launch_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.NoteApp.NoActionBar">
-        <activity android:name="project.paveltoy.noteapp.MainActivity">
+        <activity android:name="project.paveltoy.noteapp.MainActivity"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
