--- app/src/main/java/project/paveltoy/noteapp/CallbackContract.java
@@ -1,7 +1,0 @@
-package project.paveltoy.noteapp;
-
-public interface CallbackContract {
-    void saveNote(Note note);
-
-    void deleteNote(Note note);
-}
