+++ app/src/main/java/project/paveltoy/noteapp/Data/TemporaryClassNotes.java
@@ -1,0 +1,21 @@
+package project.paveltoy.noteapp.Data;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TemporaryClassNotes {
+    public static final String HEADER = "Заметка";
+    public static final String TEXT = "Текст";
+
+    public static List<Note> getNotes() {
+        List<Note> notes = new ArrayList<>();
+//        Note firstNote = new Note(
+//                Note.generateNewId(),
+//                "Добро пожаловать!",
+//                "Приветствуем Вас в программе бла-бла-бла, тут можно бла-бла-бла",
+//                Note.getCurrentDate()
+//        );
+//        notes.add(firstNote);
+        return null;
+    }
+}
