--- app/src/main/java/project/paveltoy/noteapp/NoteRepository.java
@@ -1,23 +1,0 @@
-package project.paveltoy.noteapp;
-
-public interface NoteRepository {
-    NoteRepository init(NoteSourceCallback noteSourceCallback);
-
-    void deleteNote(int position);
-
-    void deleteNote(Note note);
-
-    void editNote(int position, Note note);
-
-    void editNote(Note note);
-
-    void addNote(Note newNote);
-
-    Note getNote(int position);
-
-    int size();
-
-    int getPosition(Note note);
-
-    void insertNote(int position, Note note);
-}
