--- app/src/main/java/project/paveltoy/noteapp/NoteSourceCallback.java
@@ -1,5 +1,0 @@
-package project.paveltoy.noteapp;
-
-public interface NoteSourceCallback {
-    void initialized(NoteRepository notes);
-}
