--- app/src/main/java/project/paveltoy/noteapp/OnBackPressedListener.java
@@ -1,5 +1,0 @@
-package project.paveltoy.noteapp;
-
-public interface OnBackPressedListener {
-    void onBackPressed();
-}
