+++ app/src/main/java/project/paveltoy/noteapp/UI/CallbackContract.java
@@ -1,0 +1,9 @@
+package project.paveltoy.noteapp.UI;
+
+import project.paveltoy.noteapp.Data.Note;
+
+public interface CallbackContract {
+    void saveNote(Note note);
+
+    void deleteNote(Note note);
+}
