+++ app/src/main/java/project/paveltoy/noteapp/UI/ImportantNotesFragment.java
@@ -1,0 +1,21 @@
+package project.paveltoy.noteapp.UI;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import project.paveltoy.noteapp.R;
+
+public class ImportantNotesFragment extends Fragment {
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_important, container, false);
+    }
+}
