+++ app/src/main/java/project/paveltoy/noteapp/UI/OnBackPressedListener.java
@@ -1,0 +1,5 @@
+package project.paveltoy.noteapp.UI;
+
+public interface OnBackPressedListener {
+    void onBackPressed();
+}
